requirements.txt
groq==0.5.0
python-dotenv==1.0.0
opencv-python==4.9.0.80
SpeechRecognition==3.10.4
streamlit==1.35.0
PyPDF2==3.0.1
langchain==0.2.0
langchain_community==0.2.0
easyocr==1.7.1
cvzone==1.6.1
ultralytics==8.2.32
pyttsx3==2.90
sentence-transformers==3.0.1
faiss-cpu==1.8.0.post1
gTTs==2.5.1
pygame==2.5.2
mediapipe==0.10.14


# I did miss out and did not update the requirements.txt file at some moments, so for any module import errors you'll can just download the list i have mentioned below with their versions I have 

# Package                   Version
# ------------------------- -----------
# absl-py                   2.1.0      
# aiohttp                   3.9.3      
# aiosignal                 1.3.1
# altair                    5.2.0
# annotated-types           0.6.0
# anyio                     4.3.0
# attrs                     23.2.0
# beautifulsoup4            4.12.3
# blinker                   1.7.0
# cachetools                5.3.3
# certifi                   2024.2.2
# cffi                      1.16.0
# charset-normalizer        3.3.2
# click                     8.1.7
# colorama                  0.4.6
# comtypes                  1.4.4
# contourpy                 1.2.1
# cvzone                    1.6.1
# cycler                    0.12.1
# dataclasses-json          0.6.4
# dill                      0.3.8
# distro                    1.9.0
# docopt                    0.6.2
# easyocr                   1.7.1
# faiss-cpu                 1.8.0.post1
# filelock                  3.15.4
# flatbuffers               24.3.25
# fonttools                 4.53.0
# frozenlist                1.4.1
# fsspec                    2024.6.0
# gitdb                     4.0.11
# GitPython                 3.1.42
# greenlet                  3.0.3
# groq                      0.5.0
# gTTS                      2.5.1
# h11                       0.14.0
# httpcore                  1.0.5
# httpx                     0.27.0
# huggingface-hub           0.23.4
# idna                      3.6
# imageio                   2.34.2
# intel-openmp              2021.4.0
# jax                       0.4.30
# jaxlib                    0.4.30
# Jinja2                    3.1.3
# joblib                    1.4.2
# Js2Py                     0.74
# jsonpatch                 1.33
# jsonpointer               2.4
# jsonschema                4.21.1
# jsonschema-specifications 2023.12.1
# kiwisolver                1.4.5
# langchain                 0.1.12
# langchain-community       0.0.28
# langchain-core            0.1.32
# langchain-text-splitters  0.0.1
# langsmith                 0.1.29
# lazy_loader               0.4
# markdown-it-py            3.0.0
# MarkupSafe                2.1.5
# marshmallow               3.21.1
# matplotlib                3.9.0
# mdurl                     0.1.2
# mediapipe                 0.10.14
# mkl                       2021.4.0
# ml-dtypes                 0.4.0
# mpmath                    1.3.0
# multidict                 6.0.5
# mypy-extensions           1.0.0
# networkx                  3.3
# ninja                     1.11.1.1
# numpy                     1.26.4
# opencv-contrib-python     4.10.0.84
# opencv-python             4.9.0.80
# opencv-python-headless    4.10.0.84
# opt-einsum                3.3.0
# orjson                    3.10.5
# packaging                 23.2
# pandas                    2.2.1
# pillow                    10.3.0
# pip                       24.0
# pipwin                    0.5.2
# protobuf                  4.25.3
# psutil                    6.0.0
# py-cpuinfo                9.0.0
# pyarrow                   15.0.2
# PyAudio                   0.2.14
# pyclipper                 1.3.0.post5
# pycparser                 2.22
# pydantic                  2.6.4
# pydantic_core             2.16.3
# pydeck                    0.8.1b0
# pygame                    2.5.2
# Pygments                  2.17.2
# pyjsparser                2.7.1
# pyparsing                 3.1.2
# PyPDF2                    3.0.1
# pypiwin32                 223
# PyPrind                   2.11.3
# pySmartDL                 1.3.4
# python-bidi               0.4.2
# python-dateutil           2.9.0.post0
# python-dotenv             1.0.0
# pyttsx3                   2.90
# pytz                      2024.1
# pywin32                   306
# PyYAML                    6.0.1
# referencing               0.34.0
# regex                     2024.5.15
# requests                  2.31.0
# rich                      13.7.1
# rpds-py                   0.18.0
# safetensors               0.4.3
# scikit-image              0.24.0
# scikit-learn              1.5.0
# scipy                     1.13.1
# seaborn                   0.13.2
# sentence-transformers     3.0.1
# setuptools                65.5.0
# shapely                   2.0.4
# six                       1.16.0
# smmap                     5.0.1
# sniffio                   1.3.1
# sounddevice               0.4.7
# soupsieve                 2.5
# SpeechRecognition         3.10.4
# SQLAlchemy                2.0.28
# streamlit                 1.32.2
# sympy                     1.12.1
# tbb                       2021.13.0
# tenacity                  8.2.3
# threadpoolctl             3.5.0
# tifffile                  2024.6.18
# tokenizers                0.19.1
# toml                      0.10.2
# toolz                     0.12.1
# torch                     2.3.1
# torchvision               0.18.1
# tornado                   6.4.1
# tqdm                      4.66.4
# transformers              4.41.2
# typing_extensions         4.10.0
# typing-inspect            0.9.0
# tzdata                    2024.1
# tzlocal                   5.2
# ultralytics               8.2.32
# ultralytics-thop          2.0.0
# urllib3                   2.2.1
# watchdog                  4.0.0
# yarl                      1.9.4